requirements.txt
fastapi
uvicorn
debugpy
langchain-core<=0.2.10
langchain-community
langchain-experimental
langchainhub
gpt4all
chromadb
requests
beautifulsoup4
langchain_milvus
langchain-text-splitters
unstructured
unstructured[md]
tqdm
streamlit